Fix GCC warning regarding strcpy.
GCC's -Wsizeof-pointer-memaccess (activated by -Wall) checks calls of the form

  strncpy(dest, src, sizeof(X))

And emits a warning if X does not seem to make sense. In Relic, src = X= a constant
string and this triggers gcc's warning.

I replaced stncpy by memcpy. Because the strings being usd do not contain embedded
zeroes, this should not be a proble,