Correct the Miller-Rabin implementation
The value r was not correctly calculated (violating the invariant in the
comment, and causing tests to fail).

As per the Handbook of Applied Cryptography, the bases b that we try
out should satisfy 2 <= b <= a - 2. This condition ensures we stop in
time. In particular, when testing a prime $a$, we will note test the
base $b = a$, which fails.

We should exclude even numbers (this causes an exception, but this
exception is not caught by the test framework).