Add a direct map from field to curve
The set of admissible hash-to-field functions in the draft hash-to-curve RFC now far exceeds what's available md_xmd,
may exceed what is reasonable for Relic to implement, and will not include the customization described in 5.4.4.

As a consequence, this adds an API for accessing the map-to-curve mapping from pre-hashed bytes of the correct size.